More header improvements (separator, better contrast for progress bar)
