+++ src/renderer/components/TrackProgress/TrackProgress.module.css
@@ -16,10 +16,10 @@
 
 .trackProgress {
   display: block;
+  position: relative;
   width: 100%;
   height: 100%;
   background-color: var(--progress-bg);
-  box-shadow: inset 0 0 0 1px var(--border-color);
 }
 
 .trackRange {
@@ -27,6 +27,7 @@
   height: 100%;
   background-color: var(--main-color);
   box-shadow: inset 0 0 0 1px rgba(0 0 0 / 0.2);
+  min-width: 1px;
 }
 
 .progressTooltip {
